# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/auth": {
            "post": {
                "description": "授权用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "授权用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "auth user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.AuthUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.AuthUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.AuthUserResponse"
                        }
                    }
                }
            }
        },
        "/user/delete": {
            "delete": {
                "description": "删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "delete user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteUserResponse"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "post": {
                "description": "用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UserInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.UserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.UserInfoResponse"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "description": "分页用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "list user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.ListUsersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListUsersResponse"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.LoginResponse"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "更新信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateUserInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateUserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateUserInfoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "contract.AuthUserRequest": {
            "type": "object",
            "required": [
                "roles",
                "user_id"
            ],
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Role"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "contract.AuthUserResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                }
            }
        },
        "contract.BaseResponse": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "status_info": {
                    "$ref": "#/definitions/contract.StatusInfo"
                }
            }
        },
...
        "contract.DeleteUserRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "contract.DeleteUserResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                }
            }
        },
...
        "contract.ListUsersRequest": {
            "type": "object",
            "required": [
                "page_no",
                "page_size"
            ],
            "properties": {
                "page_no": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "contract.ListUsersResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                },
                "page_info": {
                    "$ref": "#/definitions/entity.Page"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.User"
                    }
                }
            }
        },
        "contract.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "contract.LoginResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                }
            }
        },
...
        "contract.StatusInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
...
        "contract.UpdateUserInfoRequest": {
            "type": "object",
            "required": [
                "user",
                "user_id"
            ],
            "properties": {
                "user": {
                    "$ref": "#/definitions/entity.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "contract.UpdateUserInfoResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                }
            }
        },
        "contract.UserInfoRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "contract.UserInfoResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                }
            }
        },
...
        "entity.Page": {
            "type": "object",
            "properties": {
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "entity.Role": {
            "type": "object",
            "required": [
                "role_id",
                "role_name"
            ],
            "properties": {
                "role_id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
...
        "entity.User": {
            "type": "object",
            "required": [
                "user_auth_id"
            ],
            "properties": {
                "accepted": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Role"
                    }
                },
                "school": {
                    "type": "string"
                },
                "submitted": {
                    "type": "integer"
                },
                "user_auth_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}